+++ Sources/OnboardingKit/UI/TutorialPageView.swift
@@ -18,10 +18,13 @@ import UIKit
  */
 @available(iOS 14.0, *)
 public struct TutorialPageView<PageType: TutorialPage, PageViewType: View>: View {
-
+    
     /**
      Create a tutorial page flow that renders tutorial pages
      in a page control.
+     
+     This initializer uses a standard ``Tutorial``, that has
+     a standard ``TutorialPage`` page collection.
 
      - Parameters:
        - tutorial: The tutorial to present.
@@ -30,12 +33,42 @@ public struct TutorialPageView<PageType: TutorialPage, PageViewType: View>: View
        - pageView: A function used to build a view for each page.
      */
     public init(
-        tutorial: Tutorial<PageType>,
+        tutorial: Tutorial,
+        pageIndex: Binding<Int>,
+        style: TutorialPageViewStyle = .standard,
+        @ViewBuilder pageView: @escaping PageViewBuilder
+    ) where PageType == TutorialPage {
+        self.tutorial = tutorial
+        self.pages = tutorial.pages
+        self.pageIndex = pageIndex
+        self.pageView = pageView
+        let appearance = UIPageControl.appearance()
+        appearance.currentPageIndicatorTintColor = UIColor(style.currentPageIndicatorTintColor)
+        appearance.pageIndicatorTintColor = UIColor(style.pageIndicatorTintColor)
+    }
+    
+    /**
+     Create a tutorial page flow that renders tutorial pages
+     in a page control.
+     
+     This initializer uses a ``GenericTutorial``, that has a
+     a generic ``TutorialPage`` page collection that matches
+     the page type of this view.
+
+     - Parameters:
+       - tutorial: The tutorial to present.
+       - pageIndex: The current page.
+       - style: The style to apply, by default ``TutorialPageViewStyle/standard``.
+       - pageView: A function used to build a view for each page.
+     */
+    public init(
+        tutorial: GenericTutorial<PageType>,
         pageIndex: Binding<Int>,
         style: TutorialPageViewStyle = .standard,
         @ViewBuilder pageView: @escaping PageViewBuilder
     ) {
         self.tutorial = tutorial
+        self.pages = tutorial.genericPages
         self.pageIndex = pageIndex
         self.pageView = pageView
         let appearance = UIPageControl.appearance()
@@ -46,18 +79,19 @@ public struct TutorialPageView<PageType: TutorialPage, PageViewType: View>: View
     public typealias PageViewBuilder = (PageType, TutorialPageInfo) -> PageViewType
 
     private let pageIndex: Binding<Int>
-    private let tutorial: Tutorial<PageType>
+    private let tutorial: Tutorial
+    private let pages: [PageType]
     private let pageView: PageViewBuilder
 
     public var body: some View {
         TabView(selection: pageIndex) {
-            ForEach(Array(tutorial.pages.enumerated()), id: \.offset) {
+            ForEach(Array(pages.enumerated()), id: \.offset) {
                 pageView(
                     $0.element,
                     .init(
                         pageIndex: $0.offset,
                         currentPageIndex: pageIndex.wrappedValue,
-                        totalPageCount: tutorial.pages.count
+                        totalPageCount: pages.count
                     )
                 ).tag($0.offset)
             }